refactor(ui,backend): decompose productivity god class and harden chat error handling
- Extracted `ProductivityHeader`, `ProductivityTabs`, and `ProductivityFilters` into independent components.
- Reduced `productivity.tsx` line count by adhering to SRP.
- Hardened exception handling in `service.py` by adding explicit `openai.OpenAIError` typing to previously blind catch-all exceptions, maintaining fallback integrity.
- Maintained existing styling and structural bounds using `DESIGN_TOKENS`.
