+++ backend/app/domain/chat/service.py
@@ -243,13 +243,12 @@ class ChatService:
         except TimeoutError:
             logger.warning("Timeout connecting to AI service for user=%s", user_id)
             yield "\n[[STATUS:error]]\nConnection timed out. Please try again later.\n"
-        except Exception as e:
-            if isinstance(e, (openai.APIConnectionError, openai.APITimeoutError, OSError)):
-                logger.warning("Connection error to AI service for user=%s", user_id)
-                yield "\n[[STATUS:error]]\nConnection error. Please try again later.\n"
-            else:
-                logger.exception("Unexpected error in chat stream for user=%s", user_id)
-                yield "\n[[STATUS:error]]\nAn unexpected error occurred.\n"
+        except openai.OpenAIError:
+            logger.warning("Connection error to AI service for user=%s", user_id)
+            yield "\n[[STATUS:error]]\nConnection error. Please try again later.\n"
+        except Exception:
+            logger.exception("Unexpected error in chat stream for user=%s", user_id)
+            yield "\n[[STATUS:error]]\nAn unexpected error occurred.\n"
 
     async def run_crew(
         self, user_message: str, user_id: UUID, accept_language: str | None = None
@@ -345,6 +344,8 @@ class ChatService:
             )
             if memories:
                 memory_context = "\n".join(f"- {m.content}" for m in memories)
+        except openai.OpenAIError:
+            logger.warning("Embedding generation failed for memory retrieval in room=%s, proceeding without", room_id)
         except Exception:
             logger.warning("Memory retrieval failed for room=%s, proceeding without", room_id)
 
@@ -412,6 +413,30 @@ class ChatService:
                     self.bg_service.update_room_summary_background(room_id, conversation_history)
                 )
 
+        except openai.OpenAIError:
+            logger.exception("Crew task failed due to OpenAI error in room=%s", room_id)
+            await chat_hub.broadcast_to_room(
+                room_id,
+                {
+                    "type": "agent_activity",
+                    "data": {
+                        "room_id": str(room_id),
+                        "agent_names": agent_names,
+                        "activity": "error",
+                        "detail": "AI Service Connection Error",
+                    },
+                },
+            )
+            await chat_hub.broadcast_to_room(
+                room_id,
+                {
+                    "type": "error",
+                    "data": {
+                        "message": "AI connection failed, please try again.",
+                        "room_id": str(room_id),
+                    },
+                },
+            )
         except Exception:
             logger.exception("Failed processing crew task for room=%s", room_id)
             await chat_hub.broadcast_to_room(
